+++ gemstone/core/stats.py
@@ -6,6 +6,9 @@ class DummyStatsContainer(object):
     def as_json(self):
         return {}
 
+    def after_method_call(self, *args, **kwargs):
+        pass
+
 
 class DefaultStatsContainer(object):
     def __init__(self):
@@ -31,8 +34,10 @@ class DefaultStatsContainer(object):
                 self.method_calls[method]["success_calls"] += 1
 
             method_stats = self.method_calls[method]
-            self.method_calls[method]["average_resp_time"] = (method_stats["average_resp_time"] * method_stats[
-                "total_calls"] + duration) / (method_stats["total_calls"] + 1)
+            self.method_calls[method]["average_resp_time"] = (method_stats["average_resp_time"] *
+                                                              method_stats[
+                                                                  "total_calls"] + duration) / (
+                                                             method_stats["total_calls"] + 1)
             self.method_calls[method]["total_calls"] += 1
 
     def as_json(self):